+++ rtps-r52/server.h
@@ -6,6 +6,6 @@
 #include "command.h"
 
 // Compatible with cmd_handler_t function
-int server_process(struct cmd *cmd, void *reply, size_t reply_sz);
+ssize_t server_process(struct cmd *cmd, void *reply, size_t reply_sz);
 
 #endif // SERVER_H
